requirements.txt
tensorflow==2.9.1
keras==2.9.0
opencv-python==4.11.0.86
pillow==10.4.0
numpy==2.0.2
h5py==3.14.0
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.6.1
jupyter==1.0.0
ipykernel==6.29.5
pandas==1.5.3
